% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comb}
\alias{comb}
\title{comb}
\usage{
comb(x, ...)
}
\arguments{
\item{x}{target}

\item{...}{additional args}
}
\description{
Internal combine fucntion for foreach loop used in get_payload()
}
\keyword{internal}
