% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comb_pload}
\alias{comb_pload}
\title{comb_pload}
\usage{
comb_pload(x, ...)
}
\arguments{
\item{x}{target}

\item{...}{additional args}
}
\description{
Internal combine function for foreach loop used in get_payload()
}
\keyword{internal}
