% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_payload.R
\name{etl_payload}
\alias{etl_payload}
\title{An internal function to set data types to various paylaod types.}
\usage{
etl_payload(dataset = NULL, convert = NULL, ...)
}
\arguments{
\item{dataset}{The data to be transformed.}

\item{convert}{The playload type to be transformed.}
}
\description{
An internal function to set data types to various paylaod types.
}
\keyword{internal}
