% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_ids}
\alias{game_ids}
\title{MLB Gameday IDs}
\format{A data frame with two columns}
\usage{
game_ids
}
\description{
A data frame with most available gids and the date at which they were played.
Note they are ordered from oldest game to newest game.
}
\keyword{datasets}
