% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_urls.R
\name{gameid2url}
\alias{gameid2url}
\title{Create urls from game_ids.}
\usage{
gameid2url(game_id = NULL, cluster = NULL)
}
\arguments{
\item{game_id}{A single \code{game_id} or a list of game_ids.}

\item{cluster}{A named parallel cluster produced by the \code{parallel} package.}

\item{...}{additional arguments}
}
\description{
Create urls from game_ids.
}
\examples{
\dontrun{
urls <- gameid2url(gids)
}

}
