% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payload.R
\name{get_payload}
\alias{get_payload}
\title{Get Gameday data from MLBAM.}
\usage{
get_payload(url, ...)
}
\arguments{
\item{url}{currently a url}

\item{...}{currently ignored}
}
\description{
Get Gameday data from MLBAM.
}
\examples{
\dontrun{
df <- get_payload(url)
}

}
