% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gid_date}
\alias{gid_date}
\title{Internal function to add a column of dates to a list of gids.}
\usage{
gid_date(gidlist = NULL, ...)
}
\arguments{
\item{gidlist}{A list from the internal data set \code{game_id}}

\item{...}{additional arguments.}
}
\description{
Internal function to add a column of dates to a list of gids.
}
\keyword{internal}
