% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kzone}
\alias{kzone}
\title{Internal dataset with coordinates for an average strike zone. K-zone oundreis borrowed from the book "Analyzing Baseball Data with R."}
\format{An object of class \code{data.frame} with 5 rows and 2 columns.}
\usage{
data(kzone)
}
\description{
Internal dataset with coordinates for an average strike zone. K-zone oundreis borrowed from the book "Analyzing Baseball Data with R."
}
\note{
last updated 2017-10-12
}
\keyword{internal}
