% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gids.R
\name{make_gids}
\alias{make_gids}
\title{Create game ids from dates.}
\usage{
make_gids(start = NULL, end = NULL, league = "mlb", cluster = NULL, ...)
}
\arguments{
\item{start}{A start date passed as a character in ISO 8601 format. \code{"2017-05-01"}}

\item{end}{An end date passed as a character in ISO 8601 format. \code{"2017-09-01"}}

\item{league}{The leauge to gather gids for. The default is \code{"mlb"}. Other options include \code{"aaa"} and \code{"aa"}.}

\item{cluster}{A named parallel cluster produced by the \code{doParallel} package.}

\item{...}{additional arguments}
}
\description{
Create game ids from dates.
}
\examples{
\dontrun{
urls <- make_gids(start, end)
}

}
