% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbgameday.R
\docType{package}
\name{mlbgameday}
\alias{mlbgameday}
\alias{mlbgameday-package}
\title{Tools to Extract Transform and Load  Major League Baseball Advanced Media's 'Gameday' Data}
\description{
Tools to Extract Transform and Load  Major League Baseball Advanced Media's 'Gameday' Data
}
\author{
Kris Eberwein
}
