% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload.gd_bis_boxscore}
\alias{payload.gd_bis_boxscore}
\title{An internal function for bis_boxscore payload.}
\usage{
payload.gd_bis_boxscore(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for bis_boxscore payload.
}
\keyword{internal}
