% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload.gd_game_events}
\alias{payload.gd_game_events}
\title{An internal function for game_events paylaod.}
\usage{
payload.gd_game_events(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for game_events paylaod.
}
\keyword{internal}
