% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload.gd_inning_all}
\alias{payload.gd_inning_all}
\title{An internal function for inning_all payload.}
\usage{
payload.gd_inning_all(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for inning_all payload.
}
\keyword{internal}
