% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload.gd_inning_hit}
\alias{payload.gd_inning_hit}
\title{An internal function for inning_hit payload.}
\usage{
payload.gd_inning_hit(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for inning_hit payload.
}
\keyword{internal}
