% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload.gd_linescore}
\alias{payload.gd_linescore}
\title{An internal function for linescore payload.}
\usage{
payload.gd_linescore(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for linescore payload.
}
\keyword{internal}
