% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload}
\alias{payload}
\alias{payload.default}
\alias{payload.gd_bis_boxscore}
\alias{payload.gd_game_events}
\alias{payload.gd_inning_all}
\alias{payload.gd_inning_hit}
\alias{payload.gd_linescore}
\title{Method for paylaod objects.}
\usage{
payload(urlz, ...)

\method{payload}{default}(urlz, ...)

\method{payload}{gd_bis_boxscore}(urlz, ...)

\method{payload}{gd_game_events}(urlz, ...)

\method{payload}{gd_inning_all}(urlz, ...)

\method{payload}{gd_inning_hit}(urlz, ...)

\method{payload}{gd_linescore}(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
Method for paylaod objects.
}
\examples{
\dontrun{
df <- payload(urlz)}

}
\keyword{internal}
