% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pitch_count}
\alias{pitch_count}
\title{Internal function to calculate balls and strikes for an atbat.}
\usage{
pitch_count(dat)
}
\arguments{
\item{dat}{A pitch table}

\item{...}{additional arguments.}
}
\description{
Internal function to calculate balls and strikes for an atbat.
}
\keyword{internal}
