% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_methods.R
\name{transform_pload}
\alias{transform_pload}
\alias{transform_pload.list_bis_boxscore}
\alias{transform_pload.df_game_events}
\alias{transform_pload.df_inning_hit}
\alias{transform_pload.list_inning_all}
\alias{transform_pload.list_linescore}
\alias{transform_pload.df_game}
\title{Method for payload objects.}
\usage{
transform_pload(payload_obj, ...)

\method{transform_pload}{list_bis_boxscore}(payload_obj, ...)

\method{transform_pload}{df_game_events}(payload_obj, ...)

\method{transform_pload}{df_inning_hit}(payload_obj, ...)

\method{transform_pload}{list_inning_all}(payload_obj, ...)

\method{transform_pload}{list_linescore}(payload_obj, ...)

\method{transform_pload}{df_game}(payload_obj, ...)
}
\arguments{
\item{payload_obj}{An object returned by \code{get_payload()}.}

\item{...}{additional arguments}
}
\description{
Method for payload objects.
}
\keyword{internal}
