% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_gids.R
\name{try_gids}
\alias{try_gids}
\title{Internal function to scrape gids from miniscorboard. The function returns a possibly valid gameday link.}
\usage{
try_gids(gids2check = NULL, cluster = NULL, ...)
}
\arguments{
\item{gids2check}{A list of gid dates in format \code{year_2017/month_10/day_01} as used in the \code{make_gids()} function.}

\item{cluster}{name of the cluster if one exists.}

\item{...}{Additional arguments.}
}
\description{
Internal function to scrape gids from miniscorboard. The function returns a possibly valid gameday link.
}
\keyword{internal}
