% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{umpire_ids}
\alias{umpire_ids}
\title{Internal dataset containing MLB Umpire IDs.}
\format{An object of class \code{data.frame} with 234 rows and 2 columns.}
\usage{
data(umpire_ids)
}
\description{
A character set with most player name and player ids.
}
\note{
last updated 2017-09-16
}
\keyword{internal}
