% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upperfirst}
\alias{upperfirst}
\title{Internal function to cast the first letter of a word to upper-case.}
\usage{
upperfirst(x)
}
\arguments{
\item{x}{A word or string to capitalize}

\item{...}{additional arguments.}
}
\description{
Internal function to cast the first letter of a word to upper-case.
}
\keyword{internal}
