% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urlTrue}
\alias{urlTrue}
\title{urlTrue}
\usage{
urlTrue(target)
}
\arguments{
\item{target}{url}
}
\description{
A utility function to run a tryCatch on a URL. Returns a logical TRUE / FALSE.
}
\keyword{internal}
