% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venue_ids}
\alias{venue_ids}
\title{Internal dataset containing MLB team IDs.}
\format{An object of class \code{data.frame} with 148 rows and 2 columns.}
\usage{
data(venue_ids)
}
\description{
A character set with most player name and player ids.
}
\note{
last updated 2017-09-21
}
\keyword{internal}
