% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_ids}
\alias{game_ids}
\title{Internal dataset containing MLB Gameday IDs.}
\format{An object of class \code{NULL} of length 0.}
\usage{
data(game_ids)
}
\description{
A character set with most available gids and the date at which they were played.
Note they are ordered from oldest game to newest game.
}
\note{
last updated 2017-08-23
}
\keyword{internal}
