% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_urls.R
\name{game_urls}
\alias{game_urls}
\title{Create urls from game_ids.}
\usage{
game_urls(game_id = NULL, cluster = NULL)
}
\arguments{
\item{game_id}{A single \code{game_id} or a list of game_ids.}

\item{cluster}{A named parallel cluster produced by the \code{doParallel} package.}

\item{...}{additional arguments}
}
\description{
Create urls from game_ids.
}
\examples{
\dontrun{
urls <- game_urls(gids)
}

}
