% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_urls.R
\name{game_urls}
\alias{game_urls}
\title{Create urls from game_ids.}
\usage{
game_urls(url_gids = NULL, cluster = NULL)
}
\arguments{
\item{url_gids}{A list of gid urls formatted by the \code{make_gids()} function.}

\item{cluster}{A named parallel cluster produced by the \code{doParallel} package.}

\item{...}{additional arguments}
}
\description{
Create urls from game_ids.
}
\examples{
\dontrun{
urls <- game_urls(gids)
}

}
