% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payload.R
\name{get_payload}
\alias{get_payload}
\title{Get Gameday data from MLBAM.}
\usage{
get_payload(start = NULL, end = NULL, league = "mlb",
  dataset = "inning_all", cluster = NULL, ...)
}
\arguments{
\item{start}{A start date passed as a character in ISO 8601 format. \code{"2017-05-01"}}

\item{end}{An end date passed as a character in ISO 8601 format. \code{"2017-09-01"}}

\item{league}{The leauge to gather gids for. The default is \code{"mlb"}. Other options include \code{"aaa"} and \code{"aa"}}

\item{dataset}{The dataset to be scraped. The default is "inning_all." Other options include, "inning_hit", "players",}

\item{cluster}{A named parallel cluster produced by the \code{doParallel} package.}

\item{...}{additional arguments}
}
\description{
Get Gameday data from MLBAM.
}
\examples{
\dontrun{
df <- get_payload(url)
}

}
