% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payload.R
\name{payload}
\alias{payload}
\alias{payload.default}
\alias{payload.player}
\alias{payload.pitch}
\alias{payload.inning}
\alias{payload.mini}
\title{Return data from a payload object.}
\usage{
payload(obj, ...)

\method{payload}{default}(obj, ...)

\method{payload}{player}(obj, ...)

\method{payload}{pitch}(obj, ...)

\method{payload}{inning}(obj, ...)

\method{payload}{mini}(obj, ...)
}
\arguments{
\item{obj}{An object created from a obj link}

\item{...}{currently ignored}
}
\description{
Return data from a payload object.
}
\examples{
\dontrun{
df <- payload(obj)}

}
