% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{payload}
\alias{payload}
\alias{payload.default}
\alias{payload.gd_inning_all}
\alias{payload.gd_linescore}
\title{Return data from a payload object.}
\usage{
payload(obj, ...)

\method{payload}{default}(obj, ...)

\method{payload}{gd_inning_all}(obj, ...)

\method{payload}{gd_linescore}(obj, ...)
}
\arguments{
\item{obj}{An object created from a obj link}

\item{...}{additional arguments}
}
\description{
Return data from a payload object.
}
\examples{
\dontrun{
df <- payload(obj)}

}
