% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payload.R
\name{payload}
\alias{payload}
\alias{payload.default}
\alias{payload.player}
\alias{payload.inning_all}
\alias{payload.inning_hit}
\alias{payload.mini}
\alias{payload.game_events}
\alias{payload.game}
\title{Return data from a payload object.}
\usage{
payload(obj, ...)

\method{payload}{default}(obj, ...)

\method{payload}{player}(obj, ...)

\method{payload}{inning_all}(obj, ...)

\method{payload}{inning_hit}(obj, ...)

\method{payload}{mini}(obj, ...)

\method{payload}{game_events}(obj, ...)

\method{payload}{game}(obj, ...)
}
\arguments{
\item{obj}{An object created from a obj link}

\item{...}{additional arguments}
}
\description{
Return data from a payload object.
}
\examples{
\dontrun{
df <- payload(obj)}

}
