% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gids.R
\name{validate_gids}
\alias{validate_gids}
\title{Internal function to validate gids produced by try_gids.}
\usage{
validate_gids(gidslist = NULL, league = "mlb", ...)
}
\arguments{
\item{gidslist}{a list of gids.}

\item{league}{The league to gather gids for. The default is \code{"mlb"}. Other options include \code{"aaa"} and \code{"aa"}.}

\item{...}{additional arguments}
}
\description{
Internal function to validate gids produced by try_gids.
}
\keyword{internal}
